1)升级了MBG的版本，使其支持org.mybatis.generator.plugins.UnmergeableXmlMappersPlugin插件 2)写了一点评论查询功能

+++ src/main/java/com/freeLearn/wall/mapper/FloorMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface FloorMapper {
-    int countByExample(FloorExample example);
+    long countByExample(FloorExample example);
 
     int deleteByExample(FloorExample example);
